fix: add interface for mailservice
